Improve the code organization
