+++ src/main.rs
@@ -5,69 +5,25 @@ extern crate diesel;
 #[macro_use]
 extern crate diesel_migrations;
 
-use dotenv;
-use dunce;
-use glob;
+use clap::ArgMatches;
+use std::process;
 
+mod cli;
 pub mod db;
 mod grm;
-pub mod models;
-pub mod schema;
-
-use clap::{App, Arg, ArgMatches, SubCommand};
 
 fn main() {
-    let cmd = App::new("Git Repository Manager")
-        .version("0.1.0")
-        .author("Julian Didier (theredfish)")
-        .about("Git Repository Manager (GRM) helps manage GIT repositories from your terminal.")
-        .subcommand(
-            SubCommand::with_name("add")
-                .about(
-                    "Search and add your git repositories for a given path; default path is the \
-                     current directory",
-                ).arg(
-                    Arg::with_name("path")
-                        .value_name("PATH")
-                        .help("A relative / absolute path or a directory hierarchy.")
-                        .takes_value(true),
-                ),
-        ).subcommand(
-            SubCommand::with_name("list")
-                .about("List your saved repositories for the given pattern")
-                .arg(
-                    Arg::with_name("pattern")
-                        .value_name("PATTERN")
-                        .help("The pattern to apply")
-                        .takes_value(true),
-                ),
-        ).subcommand(
-            SubCommand::with_name("goto")
-                .about("Go to the repository directory")
-                .arg(
-                    Arg::with_name("repo_name")
-                        .value_name("REPOSITORY_NAME")
-                        .help("The repository name")
-                        .takes_value(true),
-                ),
-        ).subcommand(
-            SubCommand::with_name("rm")
-                .about("Remove a git repository from the list")
-                .arg(
-                    Arg::with_name("repo_name")
-                        .value_name("REPOSITORY_NAME")
-                        .help("The repository name")
-                        .takes_value(true),
-                ),
-        ).get_matches();
+    if let Err(e) = db::run_pending_migrations() {
+        eprintln!("GRM update failed : {}", e);
+        eprintln!("Please open an issue with your output : https://github.com/theredfish/git-repo-manager/issues/new");
+        process::exit(1);
+    }
 
-    run(cmd)
+    run(cli::init());
 }
 
-fn run(cmd: ArgMatches) {
-    grm::update_db();
-
-    match cmd.subcommand() {
+fn run(matches: ArgMatches) {
+    match matches.subcommand() {
         ("add", Some(add_matches)) => {
             let git_pattern = String::from("/**/*.git");
             let add_path = add_matches.value_of("path").unwrap_or(".");
@@ -86,7 +42,7 @@ fn run(cmd: ArgMatches) {
         }
         ("", None) => {
             eprintln!("error : not enough argument. ");
-            println!("{}", cmd.usage());
+            println!("{}", matches.usage());
             println!("For more information try --help");
         }
         _ => unreachable!(),
